chore: Rename controller to end with api
